pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "erc-8004-py"
version = "0.1.0"
description = "Python SDK for ERC-8004 Trustless Agents protocol"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["erc8004", "agents", "reputation", "validation", "blockchain", "ethereum"]
authors = [
    {name = "ERC-8004 Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "web3>=6.0.0",
    "requests>=2.28.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "python-dotenv>=1.0.0",
]
brownie = [
    "eth-brownie>=1.19.0",
]

[project.urls]
Homepage = "https://github.com/tetratorus/sdk"
Documentation = "https://github.com/tetratorus/sdk"
Repository = "https://github.com/tetratorus/sdk"
"Bug Tracker" = "https://github.com/tetratorus/sdk/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
